Add use namespaced utils
