+++ packages/common/src/styles/RTLProvider.test.tsx
@@ -1,12 +1,12 @@
 import { act } from 'react-dom/test-utils';
-import { useI18N, useRtl } from '@common/intl';
+import { IntlUtils } from '@common/intl';
 import { renderHookWithProvider } from '@common/utils';
 
 describe('RTLProvider', () => {
   it('Sets the direction of the body to be rtl when a rtl language is the current locale', () => {
     const useHook = () => {
-      const isRtl = useRtl();
-      const i18n = useI18N();
+      const isRtl = IntlUtils.useRtl();
+      const i18n = IntlUtils.useI18N();
 
       return { isRtl, i18n };
     };
