+++ packages/common/src/styles/useAppTheme.ts
@@ -1,12 +1,12 @@
 import { Direction } from '@mui/material/styles';
 import { Theme } from '@mui/material';
 import { themeOptions, createTheme } from './theme';
-import { useRtl } from '@common/intl';
+import { IntlUtils } from '@common/intl';
 import { useLocalStorage } from '../localStorage';
 import merge from 'lodash/merge';
 
 export const useAppTheme = (): Theme => {
-  const isRtl = useRtl();
+  const isRtl = IntlUtils.useRtl();
   const [customTheme] = useLocalStorage('/theme/custom');
   const direction: Direction = isRtl ? 'rtl' : 'ltr';
   const rtlThemeOptions = { ...themeOptions, direction };
